+++ BackendBench/suite/__init__.py
@@ -14,23 +14,21 @@ backend implementations by comparing them against PyTorch operations.
 """
 
 from .base import OpTest, Test, TestSuite
-from .custom_ops import CustomOpsTestSuite
 from .facto import FactoTestSuite
 from .opinfo import OpInfoTestSuite
 from .smoke import randn, SmokeTestSuite
 from .torchbench import TorchBenchOpTest, TorchBenchTestSuite
-from .custom_fs import CustomFilesystemTestSuite
+from .custom_ops import CustomOpsTestSuite
 
 __all__ = [
     "Test",
     "OpTest",
     "TestSuite",
-    "CustomOpsTestSuite",
     "FactoTestSuite",
     "OpInfoTestSuite",
     "SmokeTestSuite",
     "randn",
     "TorchBenchOpTest",
     "TorchBenchTestSuite",
-    "CustomFilesystemTestSuite",
+    "CustomOpsTestSuite",
 ]
